Add files via upload
Caretakers frequently fail to administer prescriptions on time and struggle to follow consumption patterns, particularly when there are too many sick patients to care for, whether at an old age home, orphanage home, or hospital.  They can sometimes provide the wrong medications

 Care Companion

The solution to these problems was to develop an easy-to-use app to aid Caretakers in orphanages, elderly care facilities, and hospitals. This application is essentially a reminder app that notifies carers when it is time to serve medicines to the needy.
A carer can also utilize the chatbot to learn about any general medical problem, treatments, side effects, home cures, and so on.